added provision for dry run

+++ jsonconsolidator/jsonconsolidator.py
@@ -10,7 +10,7 @@ __version__ = "0.0.1"
 import sys
 from .stuff import Stuff
 import re, os, glob
-import json
+import json, copy
 
 # bids_dspath = '/Users/suyashdb/Downloads/ds212'
 
@@ -41,8 +41,22 @@ def main():
     alljson_list = []
     allfiles = []
     iter_top_flag = 0
+    change_file = False
+    iterate_topJsons = False
+    changefile_list = []
+    deletefile_list = []
+    newfile_list = []
+
+    def compareList(li):
+        result = copy.deepcopy(li[0])
+        for element in li:
+            for key in element:
+                if key in result and result[key] != element[key]:
+                    del result[key]
+        return result
+
     if not glob.glob(os.path.join(bids_dspath, 'sub-*/ses*')):
-        print("No sessions")
+        print("dataset has single session")
         search_path = os.path.join(bids_dspath,'sub-*/func/sub*task*.json')
         task_search_path = os.path.join(bids_dspath, 'sub-*/func/')
     else:
@@ -77,8 +91,9 @@ def main():
         for file in taskfiles:
             json_list.append(json.loads(open(file).read()))
         #iter all dict in the list to find common key value pair
-        common = dict(pair for pair in json_list[0].items()
-         if all((pair in d.items() for d in json_list[1:])))
+        common = compareList(json_list)
+        # common = dict(pair for pair in json_list[0].items()
+        #  if all((pair in d.items() for d in json_list[1:])))
         #create top level json for every task
         task_json_path = os.path.join(bids_dspath, ('task-%s_bold.json' % (task)))
         #write common to top level task jsons
@@ -88,8 +103,12 @@ def main():
                 common = {k:v  for k, v in old_data.items()  for k1,v1 in common.items() if k ==k1 and v==v1}
                 if common != old_data:
                     common.update(old_data)
-            with open(task_json_path, 'w') as outfile:
-                json.dump(common, outfile, indent = 4)
+            if change_file is True:
+                with open(task_json_path, 'w') as outfile:
+                    json.dump(common, outfile, indent = 4)
+            else:
+                # print(task_json_path," -file will be created")
+                newfile_list.append(task_json_path)
             iter_top_flag = iter_top_flag + 1 #count how many tasks have common k:v
 
         '''delete common elements from each task json or delete jsons if all k:v are same in all jsons '''
@@ -98,21 +117,45 @@ def main():
             value = { k : file_data[k] for k in set(file_data) - set(common) }
             '''if the whole data is common then delete the files at base level'''
             if not value:
-                print("all jsons are common hence base level json files are deleted, and a top level task json file is created")
-                os.remove(basefile)
+                if change_file is True:
+                    print("all jsons are common hence base level json files are deleted, and a top level task json file is created")
+                    os.remove(basefile)
+                else:
+                    # print(basefile, " - will be deleted")
+                    deletefile_list.append(basefile)
             else:
-                if any(isinstance(el, list) for el in value['SliceTiming']):    # checks if slicetiming is list of lists
-                    if 'SliceTiming' in value.keys():
-                        value['SliceTiming'] = sum(value['SliceTiming'], []) # flatten list of lists
-                with open(basefile, 'w') as writefile:
-                    json.dump(value, writefile, indent = 4)
+                for k, v in value.items():
+                    if type(value[k]) is list:
+                        if any(isinstance(el, list) for el in value[k]):    # checks if slicetiming is list of lists
+                            if k in value.keys():
+                                value[k] = sum(value[k], []) # flatten list of lists
+                #print("not common values are -  ", value, task)
+                if change_file is True:
+                    iterate_topJsons == True
+                    print("rewriting after removing common key:value pairs from json at subject level - ", basefile)
+                    with open(basefile, 'w') as writefile:
+                        json.dump(value, writefile, indent = 4)
+                else:
+                    # print("Dry Run ... Following file will be re-written after removing common k:v pairs - ", basefile)
+                    changefile_list.append(basefile)
+                # if any(isinstance(el, list) for el in value['SliceTiming']):    # checks if slicetiming is list of lists
+                #     if 'SliceTiming' in value.keys():
+                #         value['SliceTiming'] = sum(value['SliceTiming'], []) # flatten list of lists
+                # with open(basefile, 'w') as writefile:
+                #     json.dump(value, writefile, indent = 4)
+        print("Common k:v pairs for task-", task, '\n', common)
+        print("***** Files to be changed after removing above common k:v pair***** \n ", changefile_list, '\n')
+        changefile_list = []
+    print("**** Files that will be deleted****\n", deletefile_list, '\n')
+    print("***** New Files that will be created are *****\n", newfile_list, '\n' )
+
 
 
 
 
     #check for items common in all task jsons created at top level
     # if iterate_topJsons is True:
-    if iter_top_flag == len(set(tasks)):
+    if iter_top_flag == len(set(tasks)) & iterate_topJsons == True:
         filesearch = 'task-*bold.json'
         alltaskfiles = glob.glob(os.path.join(bids_dspath, filesearch))
         for file in alltaskfiles:
